requirements.txt
praw==7.6.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.2
textblob==0.17.1
transformers==4.31.0
tensorflow==2.17.0
tf-keras==2.17.0
A   